requirements.txt
discord.py==2.7.1
python-dotenv==1.2.2
aiohttp==3.13.3
yt-dlp==2026.3.3
PyNaCl==1.5.0
